requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
outcome==1.2.0
Pillow==10.0.1
plyer==2.1.0
pyobjus==1.2.2
PySocks==1.7.1
requests==2.31.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.22.2
trio-websocket==0.10.4
urllib3==2.0.5
wsproto==1.2.0